+++ linux/rw_svgalib.c
@@ -313,11 +313,11 @@ void SWimp_AppActivate( qboolean active )
 Sys_MakeCodeWriteable
 ================
 */
-void Sys_MakeCodeWriteable (unsigned long startaddr, unsigned long length)
+void Sys_MakeCodeWriteable (unsigned int startaddr, unsigned int length)
 {
 
 	int r;
-	unsigned long addr;
+	unsigned int addr;
 	int psize = getpagesize();
 
 	addr = (startaddr & ~(psize-1)) - psize;
